fixes bugs with sockets
