+++ views/main.html
@@ -57,8 +57,6 @@
     ></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/dayjs/1.11.3/plugin/advancedFormat.min.js"></script>
   </body>
-  <script src="/js/topics.js"></script>
   <script src="/js/main.js"></script>
-  <script src="/js/chat.js"></script>
   <script src="/js/logout.js"></script>
 </html>
